YH-2065: refactor
