+++ src/_pages/NewLandingPage/ui/NewLandingPage.tsx
@@ -1,14 +1,14 @@
 import { Flex } from '@/shared/ui/Flex';
-import { TrainerProgressBlock } from '@/widgets/Landing/TrainerProgressBlock';
 import { DifficultySection } from '@/widgets/NewLanding/DifficultySection';
 import { PrepareForInterviewSection } from '@/widgets/NewLanding/PrepareForInterviewSection';
+import { TrainerProgressSection } from '@/widgets/NewLanding/TrainerProgressSection';
 
 export const NewLandingPage = () => {
 	return (
 		<Flex direction="column" gap="80">
 			<DifficultySection />
 			<PrepareForInterviewSection />
-			<TrainerProgressBlock />
+			<TrainerProgressSection />
 		</Flex>
 	);
 };
