--- src/widgets/Landing/TrainerProgressBlock/index.ts
@@ -1,2 +1,0 @@
-export { TrainerProgressBlock } from './ui/TrainerProgressBlock/TrainerProgressBlock';
-export { TrainerProgressBlockSkeleton } from './ui/TrainerProgressBlock/TrainerProgressBlock.skeleton';
