--- src/widgets/Landing/TrainerProgressBlock/ui/InfoBlock/InfoBlock.module.css
@@ -1,128 +1,0 @@
-.info-block-wrapper {
-    display: flex;
-    flex-direction: column;
-    height: 100%;
-    border-radius: 20px;
-    background-color: var(--color-white-900);
-}
-
-.info-block {
-    display: flex;
-    flex: 1;
-    flex-direction: row;
-    
-}
-
-.info-block-main-wrapper {
-    flex: 1;
-    min-width: 0;
-}
-
-.info-block-main {
-    display: flex;
-    flex: 1;
-    flex-direction: column;
-    justify-content: space-between;
-    padding: 20px;
-    height: 100%;
-}
-
-.indicator {
-	margin: 9px;
-	background-color: var(--color-purple-700);
-}
-
-.arrow {
-    width: 24px;
-    height: 24px;
-    color: var(--color-purple-700);
-}
-
-.link {
-    display: flex;
-    align-items: center;
-    gap: 8px;
-    margin-top: auto;
-}
-
-.image-wrapper {
-    position: relative;
-    display: flex;
-    flex: 1;
-    justify-content: center;
-    align-items: center;
-    margin-top: 20px;
-    width: 100%;
-    max-width: 280px;
-    max-height: 310px;
-    box-shadow: 0 4px 10px 0 #374a011a;
-    border-radius: 14px;
-  border-bottom-left-radius: 0;
-
-}
-
-.image-wrapper img {
-    padding: 10px 0 0 10px;
-    max-width: 100%;
-    max-height: 100%;
-    object-fit: contain;
-}
-
-.desktop,
-.single {
-    display: block;
-    width: 100%;
-    height: 100%;
-    object-fit: contain;
-}
-
-.mobile {
-    display: none;
-    width: 100%;
-    height: 100%;
-    object-fit: contain;
-}
-
-@media (width <= 361px) {
-    .mobile {
-        display: block;
-    }
-
-    .desktop {
-        display: none;
-    }
-    
-    .single {
-        margin: 0 auto;
-        width: 253px;
-        height: 180px;
-    }
-    
-    .image-wrapper {
-        margin-top: 0;
-        margin-right: auto;
-        margin-left: auto;
-        width: 253px;
-        height: 180px;
-    }
-
-    .info-block {
-        display: grid;
-        grid-template-columns: 1fr;
-        gap: 16px;
-        padding: 20px;
-    }
-
-    .info-block-main-wrapper {
-        grid-row: 2;
-    }
-
-    .info-block-main {
-        padding: 0;
-        min-height: auto;
-    }
-
-    .indicator {
-        display: none;
-    }
-}
