--- src/widgets/Landing/TrainerProgressBlock/ui/TrainerProgressBlock/TrainerProgressBlock.module.css
@@ -1,43 +1,0 @@
-.trainer-progress {
-	display: flex;
-	flex-direction: column;
-	align-items: center;
-	line-height: 130%;
-	white-space: pre-line;
-}
-
-.title {
-	margin-bottom: 10px;
-    line-height: 120%;
-    font-weight: var(--font-weight-500);
-	font-size: var(--font-size-h-xs);
-}
-
-.subtitle {
-	margin-bottom: 20px;
-}
-
-.trainer-progress-wrapper {
-    display: grid;
-    grid-template-columns: 1fr 1fr;
-    align-items: stretch;
-    gap: 20px;  
-}
-
-.trainer-progress-wrapper > * {
-    height: 100%;
-}
-
-.title-block {
-	align-self: baseline;
-   
-}
-
-
-
-@media (width < 768px) {
-	.trainer-progress-wrapper {
-		grid-template-columns: 1fr;
-		margin-bottom: 70px;
-	}
-}
