+++ src/widgets/NewLanding/TrainerProgressSection/index.ts
@@ -1,0 +1,1 @@
+export { TrainerProgressSection } from './ui/TrainerProgressSection/TrainerProgressSection';
