+++ src/widgets/NewLanding/TrainerProgressSection/model/assets/index.ts
@@ -1,3 +1,2 @@
-export { default as trainer } from './trainer.png';
 export { default as trainerMobile } from './trainerMobile.png';
 export { default as statistics } from './statistics.png';
