+++ src/widgets/NewLanding/TrainerProgressSection/model/assets/statistics.png
