+++ src/widgets/NewLanding/TrainerProgressSection/model/assets/trainerMobile.png
