+++ src/widgets/NewLanding/TrainerProgressSection/ui/TrainerProgressSection/TrainerProgressSection.module.css
@@ -1,0 +1,20 @@
+.list {
+    display: grid;
+    grid-template-columns: repeat(2, 1fr);
+    align-items: stretch;
+    gap: 20px;  
+}
+
+.trainer-progress-wrapper > * {
+    height: 100%;
+}
+
+.title-block {
+	align-self: baseline;
+}
+
+@media (width < 1280px) {
+	.list {
+		grid-template-columns: 1fr;
+	}
+}
