+++ src/widgets/NewLanding/TrainerProgressSection/ui/TrainerProgressSection/TrainerProgressSection.tsx
@@ -2,29 +2,24 @@ import { useTranslations } from 'next-intl';
 
 import { Landing, i18Namespace } from '@/shared/config';
 import { Flex } from '@/shared/ui/Flex';
-import { Text } from '@/shared/ui/Text';
+import { SectionWrapper } from '@/shared/ui/SectionWrapper';
 
-import { trainer, trainerMobile } from '../../model/assets';
+import { trainerMobile } from '../../model/assets';
 import { statistics } from '../../model/assets';
 import { InfoBlock } from '../InfoBlock/InfoBlock';
-import styles from './TrainerProgressBlock.module.css';
+import styles from './TrainerProgressSection.module.css';
 
-export const TrainerProgressBlock = () => {
+export const TrainerProgressSection = () => {
 	const t = useTranslations(i18Namespace.landing);
+
 	return (
-		<section className={styles['trainer-progress']}>
-			<div className={styles['title-block']}>
-				<Text variant="head3" className={styles.title}>
-					{t(Landing.TRAINER_PROGRESS_TITLE).toUpperCase()}
-				</Text>
-				<Text variant="body3" className={styles.subtitle}>
-					{t(Landing.TRAINER_PROGRESS_SUBTITLE)}
-				</Text>
-			</div>
-			<Flex gap="20" className={styles['trainer-progress-wrapper']}>
+		<SectionWrapper
+			title={t(Landing.TRAINER_PROGRESS_TITLE).toUpperCase()}
+			subtitle={t(Landing.TRAINER_PROGRESS_SUBTITLE)}
+		>
+			<Flex gap="20" className={styles.list}>
 				<InfoBlock
-					image={trainer}
-					imageMobile={trainerMobile}
+					image={trainerMobile}
 					title={t(Landing.TRAINER_PROGRESS_BLOCK_FIRST_TITLE)}
 					text={t(Landing.TRAINER_PROGRESS_BLOCK_FIRST_DESCRIPTION)}
 					linkText={t(Landing.TRAINER_PROGRESS_BLOCK_FIRST_LINK)}
@@ -36,6 +31,6 @@ export const TrainerProgressBlock = () => {
 					linkText={t(Landing.TRAINER_PROGRESS_BLOCK_SECOND_LINK)}
 				/>
 			</Flex>
-		</section>
+		</SectionWrapper>
 	);
 };
